+++ TelegramBotAPI-core/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/InlineQueries/InlineQueryResult/InlineQueryResultGame.kt
@@ -1,6 +1,7 @@
 package com.github.insanusmokrassar.TelegramBotAPI.types.InlineQueries.InlineQueryResult
 
 import com.github.insanusmokrassar.TelegramBotAPI.types.InlineQueries.InlineQueryResult.abstracts.InlineQueryResult
+import com.github.insanusmokrassar.TelegramBotAPI.types.InlineQueryIdentifier
 import com.github.insanusmokrassar.TelegramBotAPI.types.buttons.InlineKeyboardMarkup
 import com.github.insanusmokrassar.TelegramBotAPI.types.gameShortNameField
 import kotlinx.serialization.SerialName
@@ -8,7 +9,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 class InlineQueryResultGame(
-    override val id: String,
+    override val id: InlineQueryIdentifier,
     @SerialName(gameShortNameField)
     val gameShortName: String,
     override val replyMarkup: InlineKeyboardMarkup? = null